refactor: http와 관련된 내용들을 코드스타그램의 서버와 분리
