+++ src/main/java/http/requestMessage/Method.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package http.requestMessage;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
