+++ src/main/java/http/requestMessage/Uri.java
@@ -1,9 +1,8 @@
-package webserver.request;
+package http.requestMessage;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webserver.handler.FileHandler;
-import webserver.handler.PostHandler;
 import webserver.path.PostPath;
 
 public class Uri {
