+++ src/test/java/webserver/request/MethodTest.java
@@ -2,6 +2,7 @@ package webserver.request;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import http.requestMessage.Method;
 
 import static org.assertj.core.api.Assertions.assertThat;
