+++ src/test/java/webserver/request/RequestBodyTest.java
@@ -4,6 +4,7 @@ import model.UserInfo;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import http.requestMessage.RequestBody;
 
 import java.util.Optional;
