+++ src/test/java/webserver/request/RequestTest.java
@@ -1,15 +1,13 @@
 package webserver.request;
 
 import model.UserInfo;
-import org.assertj.core.api.Assertions;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import http.requestMessage.Request;
 
 import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class RequestTest {
     private Request getRequest;
