+++ src/test/java/webserver/request/UriTest.java
@@ -3,6 +3,7 @@ package webserver.request;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import http.requestMessage.Uri;
 
 class UriTest {
