+++ src/test/java/webserver/response/MIMETest.java
@@ -3,7 +3,7 @@ package webserver.response;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.utils.MIME;
+import http.MIME;
 
 class MIMETest {
